Complete version of Log In - Log Out - Sign In
This is the release of the complete version for the creation of a profil. You can sign up by email and then access to your account, after that you can go on other page or log out in this page.
It takes in count most of common error you can have (bad login info, wrong email type etc...)

Next step, account info displaying, and email verification.